Move energy group code from utils to neutronics
These multigroup structures make more sense in the neutronics
section of the code. This moves them and updates other code
that uses them.

Along the way, I also moved some neutronics-specific settings
rules (old style) to the neutronics plugin.

I also moved some neutronics specific constants to the neutronics
package const module but imported them into __init__ for backwards
compatibility.
