+++ armi/physics/constants.py
@@ -32,8 +32,8 @@ The dpa cross sections are generated by running the following code:
 >>> from armi.materials import T92, HT9, SS316
 >>> import os
 >>> from armi import settings
->>> from armi.utils import units
->>> ANL33 = units.GROUP_STRUCTURE["ANL33"]
+>>> from armi.physics.neutronics import energyGroups
+>>> ANL33 = energyGroups.GROUP_STRUCTURE["ANL33"]
 >>> cs = settings.Settings()
 >>> spectrum = SPECTR(os.path.join(cs['DPAXSDirectoryPath'], "spectra\\twr_bol.SPECTR"))
