+++ armi/settings/settingsRules.py
@@ -29,9 +29,6 @@ in this module, which may be removed in the future.
 import re
 
 from armi import runLog
-from armi.physics.neutronics import NEUTRON
-from armi.physics.neutronics import NEUTRONGAMMA
-from armi.physics.neutronics import ALL
 
 OLD_TAGS = {
     "int": "int",
@@ -178,254 +175,6 @@ def PumpInertiaPercentChange(_cs, name, value):
         " of {}".format(newVal, value)
     )
     return {"pumpInertia": newVal}
-
-
-@include_as_rule("saveNeutronicsOutputs")
-def resetNeutronicsOutputsToSaveAfterRename(_cs, _name, value):
-    """
-    Reset the values of 'saveNeutronicsOutputs' setting after it was migrated to 'neutronicsOutputsToSave'.
-
-    Arguments
-    ---------
-    cs : setting object
-        ARMi object containing the default and user-specified settings.
-
-    name : str
-        Setting name to be modified by this rule.
-
-    value : str
-        Value of the setting identified by name.
-
-    Returns
-    -------
-    dict : dict
-        Updated setting name and or value.
-    """
-    newName = "neutronicsOutputsToSave"
-    if value == "True":
-        newValue = ALL
-    elif value == "False":
-        newValue = ""
-
-    runLog.info(
-        "The `neutronicsOutputsToSave` setting has been set to {} based on deprecated "
-        "`saveNeutronicsOutputs`.".format(newValue)
-    )
-    return {newName: newValue}
-
-
-@include_as_rule("gammaTransportActive")
-def removeGammaTransportActive(cs, _name, value):
-    """
-    Remove 'gammaTransportActive' and set values of 'globalFluxActive' for the same functionality.
-
-    Arguments
-    ---------
-    cs : setting object
-        ARMi object containing the default and user-specified settings.
-
-    name : str
-        Setting name to be modified by this rule.
-
-    value : str
-        Value of the setting identified by name.
-    """
-    if value == "True":
-        newValue = NEUTRONGAMMA
-    elif value == "False":
-        newValue = NEUTRON
-
-    cs["globalFluxActive"] = newValue
-    runLog.info(
-        "The `globalFluxActive` setting has been set to {} based on deprecated "
-        "`gammaTransportActive`.".format(newValue)
-    )
-
-
-@include_as_rule("globalFluxActive")
-def newGlobalFluxOptions(_cs, name, value):
-    """
-    Set new values of 'globalFluxActive' setting based on previous setting values.
-
-    Arguments
-    ---------
-    cs : setting object
-        ARMi object containing the default and user-specified settings.
-
-    name : str
-        Setting name to be modified by this rule.
-
-    value : str
-        Value of the setting identified by name.
-
-    Returns
-    -------
-    dict : dict
-        Updated setting name and or value.
-    """
-    if value == "True":
-        newValue = NEUTRON
-    elif value == "False":
-        newValue = ""
-    else:
-        newValue = value
-
-    if value != newValue:
-        runLog.info(
-            "The `globalFluxActive` setting has been set to {}.".format(newValue)
-        )
-    return {name: newValue}
-
-
-@include_as_rule("genXS")
-def newGenXSOptions(_cs, name, value):
-    """
-    Set new values of 'genXS' setting based on previous setting values.
-
-    Arguments
-    ---------
-    cs : setting object
-        ARMi object containing the default and user-specified settings.
-
-    name : str
-        Setting name to be modified by this rule.
-
-    value : str
-        Value of the setting identified by name.
-
-    Returns
-    -------
-    dict : dict
-        Updated setting name and or value.
-    """
-    if value == "True":
-        newValue = NEUTRON
-    elif value == "False":
-        newValue = ""
-    else:
-        newValue = value
-
-    if value != newValue:
-        runLog.info("The `genXS` setting has been set to {}.".format(newValue))
-    return {name: newValue}
-
-
-@include_as_rule("existingFIXSRC")
-def newFixedSourceOption(cs, _name, value):
-    """
-    Migrate the deprecated setting 'existingFIXSRC' to 'existingFixedSource'.
-
-    Arguments
-    ---------
-    cs : setting object
-        ARMi object containing the default and user-specified settings.
-
-    name : str
-        Setting name to be modified by this rule.
-
-    value : str
-        Value of the setting identified by name.
-
-    Returns
-    -------
-    dict : dict
-        Updated setting name and or value.
-    """
-    newName = "existingFixedSource"
-    if value == "True":
-        newValue = "FIXSRC"
-    else:
-        newValue = ""
-
-    if value != newValue:
-        runLog.info(
-            "The `existingFixedSource` setting has been set to {} based on deprecated "
-            "`existingFIXSRC`.".format(newValue)
-        )
-
-    return {newName: newValue}
-
-
-@include_as_rule("boundaries")
-def migrateNormalBCSetting(_cs, name, value):
-    """
-    Boundary setting is migrated from `Normal` to `Extrapolated`.
-    """
-    newValue = value
-    if value == "Normal":
-        newValue = "Extrapolated"
-
-    return {name: newValue}
-
-
-@include_as_rule("asymptoticExtrapolationPowerIters")
-def deprecateAsymptoticExtrapolationPowerIters(_cs, _name, _value):
-    """
-    The setting `asymptoticExtrapolationPowerIters` has been deprecated and replaced by
-    three settings to remove confusion and ensure proper use.
-
-    The three new settings are:
-        - numOuterIterPriorAsympExtrap
-        - asympExtrapOfOverRelaxCalc
-        - asympExtrapOfNodalCalc
-    """
-    runLog.error(
-        "The setting `asymptoticExtrapolationPowerIters` has been deprecated and replaced "
-        "with `numOuterIterPriorAsympExtrap`, `asympExtrapOfOverRelaxCalc`, "
-        "`asympExtrapOfNodalCalc`. Please use these settings for intended behavior."
-    )
-
-    raise ValueError(
-        "Setting `asymptoticExtrapolationPowerIters` has been deprecated. "
-        "See stdout for more details."
-    )
-
-
-@include_as_rule("groupStructure")
-def updateXSGroupStructure(cs, name, value):
-    from armi.utils import units
-
-    try:
-        units.getGroupStructure(value)
-        return {name: value}
-    except KeyError:
-        try:
-            newValue = value.upper()
-            units.getGroupStructure(newValue)
-            runLog.info(
-                "Updating the cross section group structure from {} to {}".format(
-                    value, newValue
-                )
-            )
-            return {name: newValue}
-        except KeyError:
-            runLog.info(
-                "Unable to automatically convert the `groupStructure` setting of {}. Defaulting to {}".format(
-                    value, cs.settings["groupStructure"].default
-                )
-            )
-            return {name: cs.settings["groupStructure"].default}
-
-
-def _migrateDpa(_cs, name, value):
-    newValue = value
-    if value == "dpaHT9_33":
-        newValue = "dpaHT9_ANL33_TwrBol"
-    elif value == "dpa_SS316":
-        newValue = "dpaSS316_ANL33_TwrBol"
-
-    return {name: newValue}
-
-
-@include_as_rule("gridPlateDpaXsSet")
-def migrateGridPlateDpa(_cs, name, value):
-    """Got more rigorous in dpa XS names."""
-    return _migrateDpa(_cs, name, value)
-
-
-@include_as_rule("dpaXsSet")
-def migrateDpaXs(_cs, name, value):
-    return _migrateDpa(_cs, name, value)
 
 
 @include_as_rule
