+++ doc/gallery-src/framework/run_isotxs.py
@@ -9,14 +9,14 @@ an existing binary cross section library file in :py:mod:`ISOTXS <armi.nuclearDa
 
 import matplotlib.pyplot as plt
 
-from armi.utils import units
+from armi.physics.neutronics import energyGroups
 from armi.tests import ISOAA_PATH
 from armi.nuclearDataIO import isotxs
 import armi
 
 armi.configure()
 
-gs = units.getGroupStructure("ANL33")
+gs = energyGroups.getGroupStructure("ANL33")
 lib = isotxs.readBinary(ISOAA_PATH)
 
 fe56 = lib.getNuclide("FE", "AA")
